+++ packages/mirror-tv-next/app/story/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import PageLogger from '~/components/page-logger'
+import PageLogger from '~/components/tracking/page-logger'
 import ArticleRelatedPosts from '~/components/story/article-related-posts'
 import ArticleSocialList from '~/components/story/article-social-list'
 import { Metadata } from 'next'
@@ -28,7 +28,7 @@ import type { SinglePost } from '~/graphql/query/story'
 import { SITE_URL } from '~/constants/environment-variables'
 
 import dynamic from 'next/dynamic'
-import MisoPageView from '~/components/shared/miso-pageview'
+import MisoPageView from '~/components/tracking/miso-pageview'
 import GA4SourceTracking from '~/components/story/ga4-source-tracking'
 import UiDownload from '~/components/shared/ui-download'
 import Article18Warning from '~/components/shared/article-18-warning'
