+++ packages/mirror-tv-next/app/topic/[slug]/page.tsx
@@ -21,7 +21,7 @@ import { formateHeroImage, handleMetaDesc } from '~/utils'
 import dynamic from 'next/dynamic'
 import { GPTPlaceholderDesktop } from '~/components/ads/gpt/gpt-placeholder'
 const GPTAd = dynamic(() => import('~/components/ads/gpt/gpt-ad'))
-import PageLogger from '~/components/page-logger'
+import PageLogger from '~/components/tracking/page-logger'
 
 export const revalidate = GLOBAL_CACHE_SETTING
